(Abstract) sys_framerate
